Update left-sidebar.blade.php
Script so that the nav-links in the left menu do not reload the entire page, they only reload the main "content" of the dashboard, gaining speed and giving the application a more professional style. If this option is activated, the views that you want to load in the main must have ONLY the content that you want to load in the main.
Caution! No need for @section('content') or @extends('layouts.app')

• profile/index.blade.php
<div class="container-fluid p-3">
     <div class="row justify-content-center">
         <div class="col-md-12">
             @livewire('profile')
         </div>
     </div>
</div>